requirements.txt
haystack-ai
neo4j
tqdm
openai
